requirements.txt
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.7.0
scikit-learn>=1.0.2
nltk>=3.7
pytest>=7.1
jupyter>=1.0.0
pandas>=1.5
# uncomment the following line to install pytorch and torchvision
torch==2.1.2; sys_platform != "linux" and sys_platform != "win32"
torchvision==0.16.2
torchaudio==2.1.2
transformers==4.37.1
datasets==2.14.6
spacy==3.7.2
dspy-ai==2.3.1
# dependencies for dspy-ai
openai<=0.28.1
python-dotenv
wget
